engine: Refactor physics code into a separate .cpp file
