+++ engine/src/frog/geometry/physics.hpp
@@ -16,75 +16,9 @@
 
 #include "frog/utils/debug.hpp"
 
+
 namespace frog::geo
 {
-
-
-template<typename T>
-class optimization_grid
-{
-    std::vector<std::vector<T>> grid;
-    ivec2 dim;
-    rect box;
-    vec2 chunk;
-
-    template<typename Func>
-    void for_each_around_impl(circle c, Func f)
-    {
-        using std::min, std::max;
-
-        auto inside = c.pos - box.top_left();
-        auto minim = ivec2{ int( ( inside.x() - c.radius ) / chunk.x() ),
-                            int( ( inside.y() - c.radius ) / chunk.y() ) };
-        auto maxim = ivec2{ int( ( inside.x() + c.radius ) / chunk.x() ),
-                            int( ( inside.y() + c.radius ) / chunk.y() ) };
-
-        for     (int y = max( 0, minim.y() ); y <= min( dim.y() - 1, maxim.y() ); ++y)
-            for (int x = max( 0, minim.x() ); x <= min( dim.x() - 1, maxim.x() ); ++x)
-                if (is_collision(c, rect_at(x, y)))
-                    f(x, y);
-    }
-
-public:
-    optimization_grid(ivec2 dim, rect box)
-        : grid(dim.x() * dim.y()),
-          dim(dim),
-          box(box),
-          chunk(box.size.x() / dim.x(), box.size.y() / dim.y()) {}
-
-    const auto& at(int x, int y) const { return grid.at( x + y * dim.x() ); }
-          auto& at(int x, int y)       { return grid.at( x + y * dim.x() ); }
-
-    rect rect_at(int x, int y) const
-    {
-        auto pos = box.top_left() + chunk * 0.5 + vec2{ x * chunk.x(), y * chunk.y() };
-        return rect{ pos, chunk };
-    }
-
-    void add_to_touching(circle c, T val)
-    {
-        for_each_around_impl(c, [&](int x, int y){ at(x, y).push_back(val); });
-    }
-
-    template<typename Func>
-    void for_each_around(circle c, Func f)
-    {
-        for_each_around_impl(c, [&](int x, int y)
-        {
-            for (auto& val : at(x, y))
-                f(val);
-        });
-    }
-
-    void clear()
-    {
-        for (auto& square : grid)
-        {
-            square.clear();
-        }
-    }
-};
-
 
 class soft_physics2d
 {
@@ -128,125 +62,17 @@ public:
     };
 
 private:
-    void apply_inertia(point& pt, float delta)
-    {
-        vec2 diff = pt.pos - pt.prev;
-        pt.prev = pt.pos;
-        pt.pos += diff * settings_.inertia;
-        pt.pos.y() += settings_.gravity * delta * delta;
-    }
+    void apply_inertia(point& pt, float delta);
 
-    void encapsulate(point& pt, rect rect)
-    {
-        rect.size *= 0.5;
-        pt.pos.x() = std::clamp(pt.pos.x(), rect.pos.x() - rect.size.x() + pt.radius,
-                                            rect.pos.x() + rect.size.x() - pt.radius);
-        pt.pos.y() = std::clamp(pt.pos.y(), rect.pos.y() - rect.size.y() + pt.radius,
-                                            rect.pos.y() + rect.size.y() - pt.radius);
-    }
+    void encapsulate(point& pt, rect rect);
 
-    void solve_collision(point& a, point& b)
-    {
-        vec2 dif = a.pos - b.pos;
-        float dist = dif.length();
-        float radii = a.radius + b.radius;
+    void solve_collision(point& a, point& b);
 
-        if (dist < radii)
-        {
-            // TODO: avoid division by 0
-            float ratio = (dist - radii) / (dist * (a.inv_weight * b.inv_weight));
-            a.pos -= dif * (ratio * a.inv_weight * 0.5);
-            b.pos += dif * (ratio * b.inv_weight * 0.5);
-        }
-    }
+    void solve_joint(joint j);
 
-    void solve_joint(joint j)
-    {
-        auto& a = points_[ map_points[ j.a ] ].second;
-        auto& b = points_[ map_points[ j.b ] ].second;
+    void solve_angle(angle alpha);
 
-        vec2 dif = a.pos - b.pos;
-        float dist = dif.length();
-        float desired = j.dist;
-
-        float ratio = (dist - desired) / (dist * (a.inv_weight * b.inv_weight));
-        a.pos -= dif * (ratio * a.inv_weight * 0.5);
-        b.pos += dif * (ratio * b.inv_weight * 0.5);
-    }
-
-    // TODO: there must be something wrong, doesn't seem to work for angles ~180 deg
-    // probably has to do with angles being (mod 360)
-    // probably just use geo::angle_diff
-    void solve_angle(angle alpha)
-    {
-        auto& a = points_[ map_points[ alpha.a ] ].second;
-        auto& b = points_[ map_points[ alpha.b ] ].second;
-        auto& c = points_[ map_points[ alpha.c ] ].second;
-
-        float cur = vec_angle(a.pos - b.pos, c.pos - b.pos);
-        float desired = alpha.angle;
-        float dif = angle_diff(cur, desired);
-
-        a.pos -= b.pos;
-        c.pos -= b.pos;
-
-        polar2 pol_a = to_polar(a.pos);
-        polar2 pol_c = to_polar(c.pos);
-
-        float ratio = dif / (a.inv_weight * b.inv_weight);
-        pol_a.theta -= (ratio * a.inv_weight * 0.5);
-        pol_c.theta += (ratio * b.inv_weight * 0.5);
-
-        a.pos = to_cartesian(pol_a);
-        c.pos = to_cartesian(pol_c);
-        a.pos += b.pos;
-        c.pos += b.pos;
-    }
-
-    void verlet_solve()
-    {
-        for (auto& [idx, pt] : points_)
-            apply_inertia(pt, settings_.delta);
-
-        auto grid = optimization_grid<std::pair<idx_t, point*>>(
-                                        settings_.grid_dim, settings_.universum);
-
-        for (int it = 0; it < settings_.iterations; ++it)
-        {
-            for (auto& [i, pt] : points_)
-                encapsulate(pt, settings_.universum);
-
-            grid.clear();
-
-            for (auto& [i, pt] : points_)
-                grid.add_to_touching(circle(pt.pos, pt.radius), { i, &pt });
-
-            for (auto& [i, pt] : points_)
-                grid.for_each_around(circle(pt.pos, pt.radius), [&](std::pair<idx_t, point*> other)
-                    {
-                        if (other.first > i)
-                            solve_collision(pt, *other.second);
-                    });
-
-            // for (int i = 0; i < points_.size(); ++i)
-            // {
-            //     auto& a = points_[i].second;
-            //     encapsulate(a, settings_.universum);
-            //     for (int j = i + 1; j < points_.size(); ++j)
-            //     {
-            //         auto& b = points_[j].second;
-            //         solve_collision(a, b);
-            //     }
-            // }
-
-
-            for (auto&[idx, j] : joints_)
-                solve_joint(j);
-
-            for (auto&[idx, a] : angles_)
-                solve_angle(a);
-        }
-    }
+    void verlet_solve();
 
     template<typename K, typename Map>
     static auto& map_at(const std::string& desc, Map& m, const K& key)
@@ -291,6 +117,8 @@ public:
         verlet_solve();
     }
 
+    void update_maps();
+
     const point& point_at(idx_t i) const { return points_[ map_at("soft_physics2d: points", map_points, i) ].second; }
           point& point_at(idx_t i)       { return points_[ map_at("soft_physics2d: points", map_points, i) ].second; }
 
@@ -331,23 +159,6 @@ public:
     {
         angles_.emplace_back(last, a);
         return last++;
-    }
-
-    void update_maps()
-    {
-        updated = last;
-
-        map_points.clear();
-        map_joints.clear();
-
-        map_points.reserve(points_.size());
-        map_joints.reserve(joints_.size());
-
-        for (std::size_t i = 0; i < points_.size(); ++i)
-            map_points.emplace(points_[i].first, i);
-
-        for (std::size_t i = 0; i < joints_.size(); ++i)
-            map_joints.emplace(joints_[i].first, i);
     }
 
     // TODO: Implement deletion.
